+++ CHANGELOG.md
@@ -38,3 +38,4 @@
 * Update magnetization functions and related tests to use Mesh object
 * Reduce default error tolerance of demag calc to 1.0 without loss of Bfield precision to 6 decimal places
 * Add function for computing which are surface faces on a tetrahedral mesh
+* Add Solver and CentroidMesh classes